refactor: address review feedback
- split LabelMigrationTab into component files (main, SeriesPreview, useProbeSuccessLabels)
- add LabelMode enum + LabelModeResponse type, replace mode: number usages
- collapse duplicated tag styles into a conditional dimmed class
- sentence-case the tab title and nav entry
- restore alphabetical ordering in test handler index
