+++ src/page/ConfigPageLayout/tabs/LabelMigrationTab/index.ts
@@ -1,0 +1,1 @@
+export * from './LabelMigrationTab';
