+++ src/test/handlers/index.ts
@@ -54,24 +54,24 @@ const API_ROUTES = {
   getGrafanaAlertRules,
   getHttpDashboard,
   getInstantMetrics,
+  getLabelMode,
   getLogsDS,
   getMetricsDS,
   getPromAlertRules,
   getRangeMetrics,
   getSecret,
   getSMDS,
-  getLabelMode,
   getTenant,
   getTenantCostAttributionLabels,
   getTenantLimits,
   getTenantSettings,
   installPlugin,
-  setLabelMode,
   listAlertsForCheck,
   listChecks,
   listFolders,
   listProbes,
   listSecrets,
+  setLabelMode,
   testCheck,
   updateAlertsForCheck,
   updateCheck,
